Only load the jQuery plugin if `_rollbar` is available
We don't load Rollbar in development by default, but the jQuery plugin is packaged together with other vendored scripts and always loaded.

In a more broader sense I think it's useful if library scripts take a two-staged approach where the 1. load and 2. are initialised, where the latter is up to the user. That way one can easily do e.g. `if (myApp.environment == 'production') { var someLibrary = new SomeLibrary(); }`

Moving the Plugin (and the base Rollbar js client) to such a setup would be desirable, but until then this solves the immediate problem (for the plugin).