Fixed IE9- bugs and relevant tests
I removed our custom merge/copy/log polyfills and replaces with
some community ones. Which means we now have some dependencies.

There were a few bugs in how we checked types in IE9- and how
we tested _enqueuePayload() and _log().

Lastly, I fixed a bunch of tests that were failing in IE9 and
verified that all tests now pass in IE9 on Browserstack.

@brianr