Fix overwriting in configure()
The recent change to use extend() had a bug where
we were not doing a deep merge so the notifier's
original payload was being overwritten.

This change fixes that behavior and provides a
new parameter to configure() which the user can
use to explicitly overwrite configuration values
with the options passed in.

@crisfole