Notifier Rewrite
The notifier is now generic and just does a lot of delegation to some objects that it owns or that
are registered with it. This is now an internal object which makes assumptions about it's
input/output. The idea is to have a higher level object, probably called Rollbar, which will create
a notifier along with all its friends, tie them together, and then use the provided hooks to
configure them based on their environment. Hence this is a pretty simple class because it does a lot
of delegation.