Queue
This is the queue for items to be sent to the backend. It doesn't actually queue items at first, it
just tells the API to send them. It does however handle the retry logic and manages that queue. It
also tracks requests in flight so it allows for the wait/shutdown functionality.

This uses the RateLimiter to determine whether it should send requests to the backend or not. It
also has a configurable set of predicates which can be used to stop certain items from being sent to
the backend. This is how we can implement things like ignoring items based on properties of the
payload, or ignoring items that have too low of a level, etc. Again this is configurable so that the
different execution environments can set these up how they see fit.