Fixes #290 handle objects without Object in prototype chain
Sometimes people use `Object.create(null)` which means that Object does not end up in the prototype
chain of said created objects. This means that hasOwnProperty is not a method on that object so
calling `obj.hasOwnProperty(prop)` will break. To get around this, we can use
`Object.prototype.hasOwnProperty.call(obj, prop)`. As far as I can tell this is safe at least as far
back as IE 6 and all version of node. It is only relevant when we don't really know how the object
was constructed, so it isn't necessary to use this universally. There is a miniscule performance hit
and a bit of a readability issue with it, so prefer not to use this style when we create the object
ourselves.