Fix #331 Verbose logging was using the wrong key structure
The queue gets the item to log that represents the data in the final payload structure of
```
{
  "access_token": "abc123",
  "data": { ... }
}
```

rather than the whole payload itself. The API object is responsible for wrapping that data object up
inside a structure like the above and then sending it. Thus the body of the item is a top level key
rather than nested under the data key.

Secondly, everyone wants the original error to be output if there is one isntead of just the
exception message, so make it so.