[WIP] Scrubbing for telemetry
This implements two possible additional scrubbing hooks for telemetry data:

1. Scrub all input values
2. A function which takes a description of a dom node and returns true if we should scrub the value

The first option, either scrub all values or only password fields, is the easiest to implement, is
the most straight forward, and will definitely allow people to ensure they do not leak data.

The second option, is more complicated to use but allows for maximum flexibility, it subsumes the
first option actually by just using a function that always returns true. The structure that
describes the dom node is:

```
{
  tagName,
  id,
  classes,
  attributes
}
```

It is an internal structure used to generate the string representation of an element that we embed
in the telemetry event. This allows people to say specific things about dom nodes although it does
not actually give the hierarchical information that the string provides. It somewhat depends on how
people structure their dom on what they need to be able to scrub effectively.