Update README.md
For Angular2+ projects with aot on the example would run users in to an error , arrow functions are not supported in aot.
The next issue would be the ErrorHandler , in the example it would override the existing ErrorHandler which would remove console.error , adding it. The naming of the ErrorHandler to reflect its purpose.
Next we create a wrapper service for Rollbar to exclude naming conflicts and avoid usage of the injector.
Also added some information about npm installation of the lib.