Options should override not merge
Fixes #582

We use the `node-extend` package for merging objects. This treats arrays
as objects with integer keys when it attempts to deep merge objects.
This is problematic for configuration options which are arrays because
what happens is that setting a configuration option to an array value
will just overwrite part of the array (depending on the sizes of the
arrays involved).

We do not actually require deep merging for the places that we were
using it, so instead I switch to using shallow merging for configuration
data, this allows arrays (like scrubFields) to become overwriteable.