XHR readyState and network telemetry
We do our monkey patch of the onreadystatechage handler inside send
which implies that we will never actually see readyState == 1 because of
the semantics of readyState. This change makes it so we actually start
tracking network requests sooner which means that an error while the
request is in-flight will contain telemetry data about the request. The
previous behaviour was that we only saw the request in the telemetry
data if the request completed before the error.