Concat configured scrub fields with default
The most common use case for changing the scrubFields configuration
setting is to add a small number of fields to the existing default
scrubFields. This way you automatically scrub things like "password"
but also your own custom private data like "my-auth-token". We have two
choices, we either:
 * concat your supplied list to the default
 * overwrite the default with whatever you supply
If we do the first then how do you remove things you don't want scrubbed
but which are in the default list? (This is an unlikely but possible
scenario). If we do the second, how do you add all of the defaults
without having to copypasta them from our code? (Copy paste isn't so bad
for this use case).

We could solve two by exposing the default scrub fields statically and
then you could build your own array with
`Rollbar.defaultScrubFields.concat(myFields)`.

We can solve the first problem by introducing another option
`scrubFieldsWhitelist` which is a list of keys that you do not want to
scrub even if it is also in the scrubFields list.

This PR implements this second approach, but as I write this I am
thinking exposing the static is maybe the cleaner solution.